omp_get_wall

+++ src/solvers/eri_cd_solver_class.F90
@@ -104,7 +104,7 @@ contains
 !
       real(dp), dimension(solver%n_ao,1), optional :: screening_vector
 !
-      real(dp):: s_determine_basis, e_determine_basis, s_build_vectors, e_build_vectors, omp_get_wtime
+      real(sp):: s_determine_basis, e_determine_basis, s_build_vectors, e_build_vectors, omp_get_wtime
 !
       write(output%unit, '(/a51/)') ':: Cholesky decomposition of two-electron ao_integrals'
       flush(output%unit)
